Add possibility to select arbitrary elements
Up to now, an element has to have an 'id' attribute to be able to get
an popover/tooltip. This was hardcoded in the underlying Javascript. A
new parameter was added to add[Tooltip|Popover] to override the default
behaviour to look for the element by id. With this flag set, the user
can provide an arbitrary jQuery selector as 'id'.

The argument name 'id' is maybe misleading in this case, but this
approach seemed to be minimal evasive and should not be breaking old
code.

Closes: #124